refacto(frontend) :  refacto du reste du frontend,  pour respecter les regle établie dans architecture.md (j'ai demandée a claude sonnet 4.7 pour faire le gros du travailes pour terminer la  refacto des composant
il reste les libs et le stypes  globales
