+++ src/network/projet_quizz/frontend/src/composant/molecules/QuestionsLlmImportPanel/QuestionsLlmImportPanel.tsx
@@ -4,13 +4,22 @@ import { QuestionsLlmImportPromptPanel } from "../QuestionsLlmImportPromptPanel"
 import { QUESTIONS_LLM_IMPORT_PANEL_STYLES } from "./QuestionsLlmImportPanel.styles";
 import type { QuestionsLlmImportPanelProps } from "./QuestionsLlmImportPanel.types";
 
-export function QuestionsLlmImportPanel({ options, setOptions, llmImportWorkflow }: QuestionsLlmImportPanelProps) {
-  const llmPromptFull = llmImportWorkflow.buildPrompt(options);
+export function QuestionsLlmImportPanel(props: QuestionsLlmImportPanelProps) {
+  const { data, actions } = props;
+  const { options, llmImportWorkflow } = data;
+  const prompt = llmImportWorkflow.buildPrompt(options);
+
   return (
     <Card class={QUESTIONS_LLM_IMPORT_PANEL_STYLES.card}>
       <div class={QUESTIONS_LLM_IMPORT_PANEL_STYLES.layout}>
-        <QuestionsLlmImportOptionsPanel data={{ options }} actions={{ onOptionsChange: setOptions }} />
-        <QuestionsLlmImportPromptPanel data={{ prompt: llmPromptFull }} actions={{ importFromJson: llmImportWorkflow.importFromJson }} />
+        <QuestionsLlmImportOptionsPanel
+          data={{ options }}
+          actions={{ onOptionsChange: actions.onOptionsChange }}
+        />
+        <QuestionsLlmImportPromptPanel
+          data={{ prompt }}
+          actions={{ importFromJson: llmImportWorkflow.importFromJson }}
+        />
       </div>
     </Card>
   );
