+++ src/network/projet_quizz/frontend/src/composant/molecules/QuestionsLlmImportPromptPanel/QuestionsLlmImportPromptPanel.styles.ts
@@ -6,4 +6,8 @@ export const QUESTIONS_LLM_IMPORT_PROMPT_PANEL_STYLES = {
     "textarea textarea-bordered w-full min-h-48 resize-y rounded-xl border-base-content/15 bg-base-100/90 font-mono text-xs leading-relaxed text-base-content",
   toolbar: "mb-2 flex flex-wrap items-center justify-between gap-2",
   toolbarTitle: "text-sm font-medium text-base-content",
+  pasteHelp: "mb-2 text-xs text-base-content/55",
+  jsonTextarea:
+    "textarea textarea-bordered mb-3 w-full min-h-32 rounded-2xl border-dashed border-learn/35 bg-base-100/60 font-mono text-xs leading-relaxed",
+  feedback: "mb-3 text-sm text-base-content/80",
 } as const;
