+++ src/network/projet_quizz/frontend/src/composant/organismes/CollectionsView/CollectionsView.styles.ts
@@ -1,0 +1,8 @@
+export const COLLECTIONS_VIEW_STYLES = {
+  root: "flex min-h-dvh flex-col",
+  loading: "text-sm text-base-content/60",
+  errorBox:
+    "rounded-box border border-base-content/10 bg-base-200/40 px-4 py-8 text-center text-sm text-base-content/65",
+  errorBoxMessage: "mb-3",
+  hiddenFileInput: "hidden",
+} as const;
