+++ src/network/projet_quizz/frontend/src/composant/organismes/CollectionsView/index.ts
@@ -1,1 +1,2 @@
+export * from "./CollectionsView.types";
 export * from "./CollectionsView";
