+++ src/network/projet_quizz/frontend/src/composant/organismes/DatabaseTransferView/index.ts
@@ -1,1 +1,2 @@
+export * from "./DatabaseTransferView.types";
 export * from "./DatabaseTransferView";
