+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsLlmImportCard/QuestionsLlmImportCard.hook.ts
@@ -3,24 +3,22 @@ import { importQuestionsJson } from "../../../lib/api";
 import { normalizeAndValidateImportText } from "../../../lib/llmImportNormalize";
 import { downloadCollectionAsAppJson } from "../../../lib/collectionAppJson";
 import {
-  LLM_PROMPT_BASE,
-  LLM_PROMPT_COLLECTION,
   LLM_QUESTION_COUNT_OPTIONS,
   formatExistingQuestionStemsForPrompt,
 } from "../../../lib/llmImportPrompts";
-import { QUESTION_CATEGORIE_DEFINITIONS } from "../../../lib/questionCategories";
 import { CATEGORY_OPTION_ID, type LlmImportOption } from "../../molecules/QuestionsLlmImportOptionsPanel";
 import type { LlmImportWorkflow } from "../../molecules/QuestionsLlmImportPanel";
+import {
+  COLLECTION_NAME_OPTION_ID,
+  EXISTING_STEMS_OPTION_ID,
+  QUESTION_COUNT_OPTION_ID,
+  SUBJECT_OPTION_ID,
+  buildImportSuccessMessage,
+  buildLlmImportPrompt,
+} from "./QuestionsLlmImportCard.metier";
 import type { QuestionsLlmImportCardProps } from "./QuestionsLlmImportCard.types";
+import { getOptionValue } from "./QuestionsLlmImportCard.utils";
 
-const QUESTION_COUNT_OPTION_ID = "question_count";
-const COLLECTION_NAME_OPTION_ID = "collection_name";
-const SUBJECT_OPTION_ID = "subject";
-const EXISTING_STEMS_OPTION_ID = "existing_stems";
-
-/**
- * Hook de logique pour la carte d'import LLM.
- */
 export function useQuestionsLlmImportCard(props: QuestionsLlmImportCardProps) {
   const { data, actions } = props;
   const { targetCollectionNumeric, collections, allModules, importTargetModuleId, questions } = data;
@@ -60,7 +58,10 @@ export function useQuestionsLlmImportCard(props: QuestionsLlmImportCardProps) {
         titre: "Nom de la collection",
         type: "texte",
         contenus: targetCollectionNumeric != null ? "Prerempli depuis la collection" : "Ex. Ma thematique",
-        value: switched && targetCollectionNumeric != null ? (col?.nom ?? "") : currentCollectionName || (targetCollectionNumeric != null ? (col?.nom ?? "") : ""),
+        value:
+          switched && targetCollectionNumeric != null
+            ? (col?.nom ?? "")
+            : currentCollectionName || (targetCollectionNumeric != null ? (col?.nom ?? "") : ""),
       },
       {
         id: SUBJECT_OPTION_ID,
@@ -74,7 +75,10 @@ export function useQuestionsLlmImportCard(props: QuestionsLlmImportCardProps) {
         id: EXISTING_STEMS_OPTION_ID,
         titre: "Eviter les repetitions",
         type: "case_a_cocher",
-        description: targetCollectionNumeric == null ? "Disponible lorsque tu filtres sur une collection." : "Lister les intitules deja en base.",
+        description:
+          targetCollectionNumeric == null
+            ? "Disponible lorsque tu filtres sur une collection."
+            : "Lister les intitules deja en base.",
         disabled: targetCollectionNumeric == null,
         value: targetCollectionNumeric == null ? false : (previousExistingValue ?? false),
         action: () => {
@@ -87,35 +91,19 @@ export function useQuestionsLlmImportCard(props: QuestionsLlmImportCardProps) {
   }, [targetCollectionNumeric, collections, questions]);
 
   const buildPrompt = useMemo(
-    () => (currentOptions: LlmImportOption[]) => {
-      const questionCount = getOptionValue(currentOptions, QUESTION_COUNT_OPTION_ID) || "5";
-      const category = getOptionValue(currentOptions, CATEGORY_OPTION_ID) || "histoire";
-      const collectionName = getOptionValue(currentOptions, COLLECTION_NAME_OPTION_ID).trim();
-      const subject = getOptionValue(currentOptions, SUBJECT_OPTION_ID).trim();
-      const existingStems = getOptionValue(currentOptions, EXISTING_STEMS_OPTION_ID).trim();
-      const catKey = category === "pratique" ? "pratique" : "histoire";
-
-      const countBlock = targetCollectionNumeric != null
-        ? `\n\n- Quantite: "questions" doit contenir exactement ${questionCount}.`
-        : `\n\n- Quantite: JSON doit representer exactement ${questionCount} questions au total.`;
-
-      const nameBlock = collectionName.length > 0 ? `\n\n- Nom de collection: ${collectionName}` : "";
-      const subjectBlock = subject.length > 0 ? `\n\n- Sujet:\n${subject}` : "";
-      const categorieBlock = `\n\n- Categorie: ${catKey} - ${QUESTION_CATEGORIE_DEFINITIONS[catKey]}`;
-      const existingBlock = targetCollectionNumeric != null && existingStems.length > 0 ? `\n\n- Eviter doublons:\n${existingStems}` : "";
-
-      if (targetCollectionNumeric == null) return LLM_PROMPT_BASE + countBlock + nameBlock + subjectBlock + categorieBlock;
-
-      const col = collections.find((entry) => entry.id === targetCollectionNumeric);
-      const nom = col?.nom ?? `id ${targetCollectionNumeric}`;
-      const mod = importTargetModuleId != null ? allModules.find((entry) => entry.id === importTargetModuleId) : undefined;
-
-      let tail = `\n\n- Collection active: ${nom} (${targetCollectionNumeric}).`;
-      if (mod) tail += `\n- Lien supercollection: ${mod.nom} (${mod.id}).`;
-
-      return LLM_PROMPT_COLLECTION + countBlock + nameBlock + subjectBlock + categorieBlock + existingBlock + tail;
-    },
-    [targetCollectionNumeric, collections, importTargetModuleId, allModules]
+    () => (currentOptions: LlmImportOption[]) =>
+      buildLlmImportPrompt({
+        questionCount: getOptionValue(currentOptions, QUESTION_COUNT_OPTION_ID) || "5",
+        category: getOptionValue(currentOptions, CATEGORY_OPTION_ID) || "histoire",
+        collectionName: getOptionValue(currentOptions, COLLECTION_NAME_OPTION_ID).trim(),
+        subject: getOptionValue(currentOptions, SUBJECT_OPTION_ID).trim(),
+        existingStems: getOptionValue(currentOptions, EXISTING_STEMS_OPTION_ID).trim(),
+        targetCollectionNumeric,
+        collections,
+        importTargetModuleId,
+        allModules,
+      }),
+    [targetCollectionNumeric, collections, importTargetModuleId, allModules],
   );
 
   const llmImportWorkflow = useMemo<LlmImportWorkflow>(
@@ -128,13 +116,10 @@ export function useQuestionsLlmImportCard(props: QuestionsLlmImportCardProps) {
         const categorieApi = getOptionValue(options, CATEGORY_OPTION_ID) === "pratique" ? "pratique" : "histoire";
         const res = await importQuestionsJson(dataJson, { collectionId: cid, moduleId: mid, categorie: categorieApi });
         actions.onImportSuccess();
-        if (cid != null) return `Import reussi: ${res.createdQuestions} question(s) creee(s).`;
-        return res.createdCollections > 0
-          ? `Import reussi: ${res.createdQuestions} question(s), ${res.createdCollections} collection(s).`
-          : `Import reussi: ${res.createdQuestions} question(s).`;
+        return buildImportSuccessMessage(res, cid != null);
       },
     }),
-    [buildPrompt, options, targetCollectionNumeric, importTargetModuleId, actions.onImportSuccess]
+    [buildPrompt, options, targetCollectionNumeric, importTargetModuleId, actions],
   );
 
   const handleExportCollectionJson = () => {
@@ -168,12 +153,4 @@ export function useQuestionsLlmImportCard(props: QuestionsLlmImportCardProps) {
       toggleImport,
     },
   };
-}
-
-/**
- * Récupère la valeur d'une option par son ID.
- */
-function getOptionValue(options: LlmImportOption[], id: string): string {
-  const option = options.find((entry) => entry.id === id);
-  return typeof option?.value === "string" || typeof option?.value === "number" ? String(option.value) : "";
 }
