+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsLlmImportCard/QuestionsLlmImportCard.utils.ts
@@ -1,0 +1,9 @@
+import type { LlmImportOption } from "../../molecules/QuestionsLlmImportOptionsPanel";
+
+export function getOptionValue(options: LlmImportOption[], id: string): string {
+  const option = options.find((entry) => entry.id === id);
+  if (typeof option?.value === "string" || typeof option?.value === "number") {
+    return String(option.value);
+  }
+  return "";
+}
