+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsLlmImportCard/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuestionsLlmImportCard.types";
 export * from "./QuestionsLlmImportCard";
