+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsTable/QuestionsTable.types.ts
@@ -1,1 +1,14 @@
-export type QuestionsTableTypesPlaceholder = Record<string, never>;
+import type { QuizzQuestionRow } from "../../../types/quizz";
+
+export type QuestionsTableProps = {
+  data: {
+    questions: QuizzQuestionRow[];
+  };
+  actions: {
+    onEdit: (question: QuizzQuestionRow) => void;
+    onRemove: (id: number) => void;
+  };
+  status: {
+    saving: boolean;
+  };
+};
