+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsTable/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuestionsTable.types";
 export * from "./QuestionsTable";
