+++ src/network/projet_quizz/frontend/src/composant/organismes/QuizResultsView/QuizResultsView.tsx
@@ -1,32 +1,22 @@
-import { useMemo } from "preact/hooks";
-import { route } from "preact-router";
 import { Trophy } from "lucide-preact";
-import {
-  buildPlaySessionQuery,
-  isPlayOrder,
-  playOrdersRequireUserId,
-  type PlayOrder,
-} from "../../../lib/playOrder";
-import { readLastQuizResult } from "../../../lib/lastQuizResult";
-import { useUserSession } from "../../../lib/userSession";
 import { AppHeader } from "../../molecules/AppHeader/AppHeader";
 import { AppFooter } from "../../molecules/AppFooter/AppFooter";
 import { Card } from "../../atomes/Card/Card";
 import { Button } from "../../atomes/Button/Button";
 import { Badge } from "../../atomes/Badge/Badge";
+import { useQuizResultsView } from "./QuizResultsView.hook";
 import { QUIZ_RESULTS_VIEW_STYLES } from "./QuizResultsView.styles";
 
 export function QuizResultsView() {
-  const { userId } = useUserSession();
-  const result = useMemo(() => readLastQuizResult(), []);
+  const { result, summary, navigation } = useQuizResultsView();
 
-  if (!result) {
+  if (result == null || summary == null) {
     return (
       <div class={QUIZ_RESULTS_VIEW_STYLES.root}>
         <AppHeader />
         <main class="fl-page-enter mx-auto flex max-w-lg flex-1 flex-col items-center justify-center gap-4 px-4 py-12 text-center">
           <p class="text-base text-base-content/70">Aucun resultat recent. Lance un quiz depuis l accueil ou une collection.</p>
-          <Button variant="flow" onClick={() => route("/")}>
+          <Button variant="flow" onClick={navigation.goHome}>
             Accueil
           </Button>
         </main>
@@ -34,25 +24,6 @@ export function QuizResultsView() {
       </div>
     );
   }
-
-  const pct = result.total <= 0 ? 0 : Math.round((result.good / result.total) * 100);
-  const replay = () => {
-    const orders: PlayOrder[] =
-      result.playOrders != null && result.playOrders.length > 0
-        ? result.playOrders
-        : result.playOrder != null && isPlayOrder(result.playOrder)
-          ? [result.playOrder]
-          : ["random"];
-    const q = buildPlaySessionQuery({
-      orders,
-      qtype: result.playQtype ?? "melanger",
-      infinite: result.playInfinite === true,
-      userId: playOrdersRequireUserId(orders) ? userId : undefined,
-    });
-    if (result.mode === "random") route(`/play/random${q}`);
-    else if (result.collectionId != null) route(`/play/${result.collectionId}${q}`);
-    else route("/collections");
-  };
 
   return (
     <div class={QUIZ_RESULTS_VIEW_STYLES.root}>
@@ -66,17 +37,19 @@ export function QuizResultsView() {
             Session terminee
           </Badge>
           <h1 class="mb-2 text-2xl font-bold tracking-tight text-base-content">Resultat</h1>
-          <p class="mb-6 text-sm text-base-content/65">{result.collectionName}</p>
-          <p class="mb-1 text-4xl font-semibold text-flow">{result.good}/{result.total}</p>
-          <p class="mb-8 text-sm text-base-content/55">{pct}% de bonnes reponses</p>
+          <p class="mb-6 text-sm text-base-content/65">{summary.collectionName}</p>
+          <p class="mb-1 text-4xl font-semibold text-flow">
+            {summary.good}/{summary.total}
+          </p>
+          <p class="mb-8 text-sm text-base-content/55">{summary.percent}% de bonnes reponses</p>
           <div class="flex flex-col gap-2 sm:flex-row sm:justify-center">
-            <Button variant="flow" onClick={replay}>
+            <Button variant="flow" onClick={navigation.replay}>
               Rejouer
             </Button>
-            <Button variant="outline" onClick={() => route("/collections")}>
+            <Button variant="outline" onClick={navigation.goCollections}>
               Collections
             </Button>
-            <Button variant="ghost" onClick={() => route("/")}>
+            <Button variant="ghost" onClick={navigation.goHome}>
               Accueil
             </Button>
           </div>
