+++ src/network/projet_quizz/frontend/src/composant/organismes/QuizResultsView/QuizResultsView.types.ts
@@ -1,1 +1,12 @@
-export type QuizResultsViewTypesPlaceholder = Record<string, never>;
+import type { LastQuizResult } from "../../../lib/lastQuizResult";
+
+export type QuizResultsViewProps = Record<string, never>;
+
+export type QuizResultsSummary = {
+  good: number;
+  total: number;
+  percent: number;
+  collectionName: string;
+};
+
+export type { LastQuizResult };
