+++ src/network/projet_quizz/frontend/src/composant/organismes/QuizResultsView/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuizResultsView.types";
 export * from "./QuizResultsView";
