+++ src/network/projet_quizz/frontend/src/composant/organismes/QuizSessionView/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuizSessionView.types";
 export * from "./QuizSessionView";
