+++ src/network/projet_quizz/frontend/src/composant/organismes/SessionDetailsView/SessionDetailsView.types.ts
@@ -1,1 +1,11 @@
-export type SessionDetailsViewTypesPlaceholder = Record<string, never>;
+import type { SessionDetail } from "../../../types/quizz";
+
+export type SessionDetailsViewProps = {
+  sessionId?: string;
+};
+
+export type SessionDetailsState = {
+  session: SessionDetail | null;
+  loading: boolean;
+  notFound: boolean;
+};
