+++ src/network/projet_quizz/frontend/src/composant/organismes/SessionDetailsView/index.ts
@@ -1,1 +1,2 @@
+export * from "./SessionDetailsView.types";
 export * from "./SessionDetailsView";
